+++ CHANGES.md
@@ -1,6 +1,12 @@
 <!-- markdownlint-configure-file { "no-duplicate-heading": { "siblings_only": true } } -->
 
 <!-- markdownlint-disable-next-line first-line-h1 -->
+## Unreleased
+
+### Fixed
+
+* Restored `ToComputed<Resource<Rc<T>>>` implementation for `LazyCache<T>`.
+
 ## 0.8.0 - 2025-07-15
 
 ### Added